applies drag functionality for improved ux
